// ordergroove/types/paths/orders/{order_id}/skip_subscription.types.ts
// This code was automatically generated from an OpenAPI description.
// Do not edit this file. Edit the OpenAPI file instead.
// For more information, see https://github.com/pmcelhaney/counterfact/blob/main/docs/faq.md

import type { WideOperationArgument } from "../../../../counterfact-types/index.ts";
import type { OmitValueWhenNever } from "../../../../counterfact-types/index.ts";
import type { MaybePromise } from "../../../../counterfact-types/index.ts";
import type { COUNTERFACT_RESPONSE } from "../../../../counterfact-types/index.ts";
import type { Context } from "../../../../routes/_.context.ts";
import type { ResponseBuilderFactory } from "../../../../counterfact-types/index.ts";
import type { Order } from "../../../components/schemas/Order.js";
import type { BadRequest } from "../../../#/components/responses/BadRequest.js";
import type { AuthenticationFailed } from "../../../#/components/responses/AuthenticationFailed.js";
import type { NotFound } from "../../../#/components/responses/NotFound.js";

/**
 * Removes all items belonging to the subscription from the order and generates
 * the next order for those items. Source reviewed 2026-08-02:
 * https://developer.ordergroove.com/reference/skip-subscription
 *
 */
export type skipSubscription = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: skipSubscription_Path;
    headers: skipSubscription_Headers;
    cookie: never;
    body: { subscription: string };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Order;
          };
        };
        examples: {};
      };
      400: BadRequest;
      403: AuthenticationFailed;
      404: NotFound;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type skipSubscription_Path = { order_id: string };

export type skipSubscription_Headers = {
  "x-api-key": string;
  Authorization: string;
};